moved test toolkit into utilities and added vm-modules ml integration test
