--- libs/vm-modules/tests/data/boston/boston_10_label.csv
@@ -1,10 +1,0 @@
-24.0;
-21.6;
-34.7;
-33.4;
-36.2;
-28.7;
-22.9;
-27.1;
-16.5;
-18.9;
