+++ libs/vm-modules/tests/integration/ml/boston_house_prices.cpp
@@ -16,16 +16,13 @@
 //
 //------------------------------------------------------------------------------
 
-#include "../unit/vm_test_toolkit.hpp"
-
 #include "vm_modules/math/tensor/tensor.hpp"
 #include "vm_modules/math/type.hpp"
-//#include "vm_modules/ml/dataloaders/dataloader.hpp"
+#include "vm_modules/test_utilities/vm_test_toolkit.hpp"
 
 #include "gmock/gmock.h"
 
 #include <regex>
-//#include <sstream>
 
 using namespace fetch::vm;
 
@@ -36,28 +33,26 @@ using DataType = fetch::vm_modules::math::DataType;
 std::string const BOSTON_HOUSING_SOURCE = R"(
     function main()
 
-//      // read in training and test data
-//      var data = readCSV("../data/boston/boston_10_data.csv");
-//      var label = readCSV("../data/boston/boston_10_label.csv");
-//
-//      // set up a model architecture
-//      var model = Model("sequential");
-//      model.add("dense", 13u64, 10u64, "relu");
-//      model.add("dense", 10u64, 10u64, "relu");
-//      model.add("dense", 10u64, 1u64);
-//      model.compile("mse", "adam");
-//
-//      // train the model
-//      var batch_size = 10u64;
-//      model.fit(data, label, batch_size);
-//
-//      // evaluate performance
-//      var loss = model.evaluate();
-//      print(loss[0u64]);
-//
-//      // make predictions on data
-//      var predictions = model.predict(data);
-//      print(predictions.at(0u64, 0u64));
+      // read in training and test data
+      var data = Tensor('0.00632,18.0;2.31,0.0;0.538,6.575;65.2,4.09;1.0,296.0;15.3,396.9;4.98,0.00632;18.0,2.31;0.0,0.538;6.575,65.2;4.09,1.0;296.0,15.3;396.9,4.98;');
+      var label = Tensor('24.0,24.0;');
+
+      // set up a model architecture
+      var model = Model("sequential");
+      model.add("dense", 13u64, 10u64, "relu");
+      model.add("dense", 10u64, 10u64, "relu");
+      model.add("dense", 10u64, 1u64);
+      model.compile("mse", "adam");
+
+      // train the model
+      var batch_size = 2u64;
+      model.fit(data, label, batch_size);
+
+      // evaluate performance
+      var loss = model.evaluate();
+
+      // make predictions on data
+      var predictions = model.predict(data);
 
     endfunction
   )";
@@ -71,31 +66,20 @@ class VMBostonTests : public ::testing::Test
 public:
   std::stringstream stdout;
   VmTestToolkit     toolkit{&stdout};
-
-  VMPtr vm;
-
-  void SetUp() override
-  {
-    // setup the VM
-    using VMFactory = fetch::vm_modules::VMFactory;
-    auto module     = VMFactory::GetModule(fetch::vm_modules::VMFactory::USE_ALL);
-    vm              = std::make_shared<fetch::vm::VM>(module.get());
-  }
-
 };
 
 TEST_F(VMBostonTests, model_add_dense_noact)
 {
-    std::string const src = BOSTON_HOUSING_SOURCE;
-    ASSERT_TRUE(toolkit.Compile(src));
-    if (IGNORE_CHARGE_ESTIMATION)
-    {
-      ASSERT_TRUE(toolkit.Run(nullptr, ChargeAmount{0}));
-    }
-    else
-    {
-      ASSERT_TRUE(toolkit.Run());
-    }
+  std::string const src = BOSTON_HOUSING_SOURCE;
+  ASSERT_TRUE(toolkit.Compile(src));
+  if (IGNORE_CHARGE_ESTIMATION)
+  {
+    ASSERT_TRUE(toolkit.Run(nullptr, ChargeAmount{0}));
+  }
+  else
+  {
+    ASSERT_TRUE(toolkit.Run());
+  }
 }
 
 }  // namespace
