+++ libs/vm-modules/tests/unit/array_tests.cpp
@@ -16,7 +16,7 @@
 //
 //------------------------------------------------------------------------------
 
-#include "vm_test_toolkit.hpp"
+#include "vm_modules/test_utilities/vm_test_toolkit.hpp"
 
 #include "gmock/gmock.h"
