+++ libs/vm-modules/tests/unit/core_language/core_etch_tests.cpp
@@ -16,9 +16,8 @@
 //
 //------------------------------------------------------------------------------
 
-#include "vm_test_toolkit.hpp"
-
 #include "core/string/replace.hpp"
+#include "vm_modules/test_utilities/vm_test_toolkit.hpp"
 
 #include "gmock/gmock.h"
