+++ libs/vm-modules/tests/unit/math/big_number_wrapper_tests.cpp
@@ -16,12 +16,11 @@
 //
 //------------------------------------------------------------------------------
 
-#include "vm_test_toolkit.hpp"
-
 #include "math/standard_functions/log.hpp"
 #include "vm_modules/math/bignumber.hpp"
 #include "vm_modules/math/math.hpp"
 #include "vm_modules/math/type.hpp"
+#include "vm_modules/test_utilities/vm_test_toolkit.hpp"
 
 #include "gmock/gmock.h"
