+++ libs/vm-modules/tests/unit/math/math.cpp
@@ -16,8 +16,6 @@
 //
 //------------------------------------------------------------------------------
 
-#include "vm_test_toolkit.hpp"
-
 #include "math/standard_functions/abs.hpp"
 #include "math/standard_functions/exp.hpp"
 #include "math/standard_functions/log.hpp"
@@ -26,6 +24,7 @@
 #include "vm/fixed.hpp"
 #include "vm_modules/math/math.hpp"
 #include "vm_modules/math/type.hpp"
+#include "vm_modules/test_utilities/vm_test_toolkit.hpp"
 
 #include "gmock/gmock.h"
