+++ libs/vm-modules/tests/unit/state_tests.cpp
@@ -16,14 +16,13 @@
 //
 //------------------------------------------------------------------------------
 
-#include "vm_test_toolkit.hpp"
-
 #include "vm/address.hpp"
 #include "vm/array.hpp"
 #include "vm/fixed.hpp"
 #include "vm/map.hpp"
 #include "vm/pair.hpp"
 #include "vm_modules/core/byte_array_wrapper.hpp"
+#include "vm_modules/test_utilities/vm_test_toolkit.hpp"
 
 #include "gmock/gmock.h"
