+++ libs/vm-modules/tests/unit/vm_charge_tests.cpp
@@ -16,10 +16,9 @@
 //
 //------------------------------------------------------------------------------
 
-#include "vm_test_toolkit.hpp"
-
 #include "vm/array.hpp"
 #include "vm/variant.hpp"
+#include "vm_modules/test_utilities/vm_test_toolkit.hpp"
 
 #include "gtest/gtest.h"
